example entries / db set up
DB is now set up to correctly name the "welcome" example entries.
In the data itself, it's slightly ugly: there are "empty" rows with unique identifiers still present. This should have no bearing on the function of the app, and will possibly be cleared by a brand new db implementation in the near future.